Move IAC stub and tweak logger
